Moved Pagination to src\Components; fixed pathing; fixed svg name format; ReactMD in Discover
