+++ src/Components/Pagination/Pagination.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactPaginate from 'react-paginate';
-import LeftButtonIcon from '../res/Left Button.svg';
-import RightButtonIcon from '../res/Right Button.svg';
-import './ResultTableMap/ResultTableMap.css'
+import LeftButtonIcon from './res/leftbutton.svg';
+import RightButtonIcon from './res/rightbutton.svg';
+import '../../Pages/ResearchingNWC/Components/ResultTableMap/ResultTableMap.css'
 
 export const Pagination = (props) => {
     return (
