+++ src/Pages/OralHistories/components/ParticipantsTable.js
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
-import leftIcon from '../../ResearchingNWC/res/Left Button.svg'
-import rightIcon from '../../ResearchingNWC/res/Right Button.svg'
+import leftIcon from '../../../Components/Pagination/res/leftbutton.svg'
+import rightIcon from '../../../Components/Pagination/res/rightbutton.svg'
 import '../../Discover/Discover.css'
 import stateTerritories from '../../../assets/stateTerritories.json';
-import { Pagination } from '../../ResearchingNWC/Components/Pagination'
+import { Pagination } from '../../../Components/Pagination/Pagination'
 import { Stack } from "../../../Components/Stack";
 import { Search } from "../../../Components/SearchBox/Search"
 import { Typography } from "../../../Components/Typography"
